[MOD] <mrp_product_variants> It creates the product.product when the MO is confirmed
