+++ mrp_product_variants/models/mrp_bom.py
@@ -218,237 +218,3 @@ class MrpBom(models.Model):
         self._get_workorder_operations(result2, level=level,
                                        routing_id=routing_id)
         return result, result2
-
-
-class MrpProductionAttribute(models.Model):
-    _name = 'mrp.production.attribute'
-
-    mrp_production = fields.Many2one(comodel_name='mrp.production',
-                                     string='Manufacturing Order')
-    attribute = fields.Many2one(comodel_name='product.attribute',
-                                string='Attribute')
-    value = fields.Many2one(comodel_name='product.attribute.value',
-                            domain="[('attribute_id', '=', attribute),"
-                            "('id', 'in', possible_values[0][2])]",
-                            string='Value')
-    possible_values = fields.Many2many(
-        comodel_name='product.attribute.value',
-        compute='_get_possible_attribute_values')
-
-    @api.one
-    @api.depends('attribute', 'mrp_production.product_template',
-                 'mrp_production.product_template.attribute_line_ids')
-    def _get_possible_attribute_values(self):
-        attr_values = self.env['product.attribute.value']
-        template = self.mrp_production.product_template
-        for attr_line in template.attribute_line_ids:
-            if attr_line.attribute_id.id == self.attribute.id:
-                attr_values |= attr_line.value_ids
-        self.possible_values = attr_values.sorted()
-
-
-class MrpProduction(models.Model):
-    _inherit = 'mrp.production'
-
-    product_id = fields.Many2one(required=False)
-    product_template = fields.Many2one(comodel_name='product.template',
-                                       string='Product', readonly=True,
-                                       states={'draft': [('readonly', False)]})
-    product_attributes = fields.One2many(
-        comodel_name='mrp.production.attribute', inverse_name='mrp_production',
-        string='Product attributes', copyable=True, readonly=True,
-        states={'draft': [('readonly', False)]},)
-
-    def product_id_change(self, cr, uid, ids, product_id, product_qty=0,
-                          context=None):
-        result = super(MrpProduction, self).product_id_change(
-            cr, uid, ids, product_id, product_qty=product_qty, context=context)
-        product_obj = self.pool['product.product']
-        product = product_obj.browse(cr, uid, product_id, context=context)
-        result['value'].update(
-            {'product_template': product.product_tmpl_id.id,
-             'product_attributes': (
-                 product._get_product_attributes_values_dict())})
-        return result
-
-    @api.multi
-    @api.onchange('product_template')
-    def onchange_product_template(self):
-        self.ensure_one()
-        if self.product_template:
-            self.product_uom = self.product_template.uom_id
-            self.product_attributes = (
-                self.product_template._get_product_attributes_dict())
-            if not self.product_template.attribute_line_ids:
-                self.product_id = (
-                    self.product_template.product_variant_ids and
-                    self.product_template.product_variant_ids[0])
-                self.product_attributes = (
-                    self.product_id._get_product_attributes_values_dict())
-            self.bom_id = self.env['mrp.bom']._bom_find(
-                product_tmpl_id=self.product_template.id)
-            self.routing_id = self.bom_id.routing_id
-            return {'domain': {'product_id':
-                               [('product_tmpl_id', '=',
-                                 self.product_template.id)],
-                               'bom_id':
-                               [('product_tmpl_id', '=',
-                                 self.product_template.id)]}}
-        return {'domain': {}}
-
-    @api.one
-    @api.onchange('product_attributes')
-    def onchange_product_attributes(self):
-        product_obj = self.env['product.product']
-        self.product_id = product_obj._product_find(self.product_template,
-                                                    self.product_attributes)
-
-    @api.multi
-    def _action_compute_lines(self, properties=None):
-        """ Compute product_lines and workcenter_lines from BoM structure
-        @return: product_lines
-        """
-        if properties is None:
-            properties = []
-        results = []
-        bom_obj = self.env['mrp.bom']
-        uom_obj = self.env['product.uom']
-        prod_line_obj = self.env['mrp.production.product.line']
-        workcenter_line_obj = self.env['mrp.production.workcenter.line']
-        for production in self:
-            #  unlink product_lines
-            production.product_lines.unlink()
-            #  unlink workcenter_lines
-            production.workcenter_lines.unlink()
-            #  search BoM structure and route
-            bom_point = production.bom_id
-            bom_id = production.bom_id.id
-            if not bom_point:
-                if not production.product_id:
-                    bom_obj._bom_find(
-                        product_tmpl_id=production.product_template.id,
-                        properties=properties)
-                else:
-                    bom_id = bom_obj._bom_find(
-                        product_id=production.product_id.id,
-                        properties=properties)
-                if bom_id:
-                    bom_point = bom_obj.browse(bom_id)
-                    routing_id = bom_point.routing_id.id or False
-                    self.write({'bom_id': bom_id, 'routing_id': routing_id})
-
-            if not bom_id:
-                raise exceptions.Warning(
-                    _('Error! Cannot find a bill of material for this'
-                      ' product.'))
-
-            # get components and workcenter_lines from BoM structure
-            factor = uom_obj._compute_qty(production.product_uom.id,
-                                          production.product_qty,
-                                          bom_point.product_uom.id)
-            # product_lines, workcenter_lines
-            results, results2 = bom_obj._bom_explode(
-                bom_point, production.product_id,
-                factor / bom_point.product_qty, properties,
-                routing_id=production.routing_id.id, production=production)
-
-            #  reset product_lines in production order
-            for line in results:
-                line['production_id'] = production.id
-                prod_line_obj.create(line)
-
-            #  reset workcenter_lines in production order
-            for line in results2:
-                line['production_id'] = production.id
-                workcenter_line_obj.create(line)
-
-            self._get_workorder_in_product_lines(self.workcenter_lines,
-                                                 self.product_lines)
-        return results
-
-    def _get_workorder_in_product_lines(self, workcenter_lines, product_lines):
-        for p_line in product_lines:
-            for bom_line in self.bom_id.bom_line_ids:
-                if ((bom_line.product_template.id == p_line.product_template.id
-                        or bom_line.product_id.product_tmpl_id.id ==
-                        p_line.product_template.id) and
-                        (not bom_line.product_id or
-                         bom_line.product_id.id == p_line.product_id.id)):
-                    for wc_line in workcenter_lines:
-                        if wc_line.routing_wc_line == bom_line.operation:
-                            p_line.work_order = wc_line
-                            break
-
-
-class MrpProductionProductLineAttribute(models.Model):
-    _name = 'mrp.production.product.line.attribute'
-
-    product_line = fields.Many2one(
-        comodel_name='mrp.production.product.line',
-        string='Product line')
-    attribute = fields.Many2one(comodel_name='product.attribute',
-                                string='Attribute')
-    value = fields.Many2one(comodel_name='product.attribute.value',
-                            domain="[('attribute_id', '=', attribute),"
-                            "('id', 'in', possible_values[0][2])]",
-                            string='Value')
-    possible_values = fields.Many2many(
-        comodel_name='product.attribute.value',
-        compute='_get_possible_attribute_values')
-
-    @api.one
-    def _get_parent_value(self):
-        if self.attribute.parent_inherited:
-            production = self.product_line.production_id
-            for attr_line in production.product_attributes:
-                if attr_line.attribute == self.attribute:
-                    self.value = attr_line.value
-
-    @api.one
-    @api.depends('attribute')
-    def _get_possible_attribute_values(self):
-        attr_values = self.env['product.attribute.value']
-        template = self.product_line.product_template
-        for attr_line in template.attribute_line_ids:
-            if attr_line.attribute_id.id == self.attribute.id:
-                attr_values |= attr_line.value_ids
-        self.possible_values = attr_values.sorted()
-
-
-class MrpProductionProductLine(models.Model):
-    _inherit = 'mrp.production.product.line'
-
-    product_id = fields.Many2one(required=False)
-    product_template = fields.Many2one(comodel_name='product.template',
-                                       string='Product')
-    product_attributes = fields.One2many(
-        comodel_name='mrp.production.product.line.attribute',
-        inverse_name='product_line', string='Product attributes',
-        copyable=True)
-
-    @api.one
-    @api.onchange('product_template')
-    def onchange_product_template(self):
-        if self.product_template:
-            product_id = self.env['product.product']
-            if not self.product_template.attribute_line_ids:
-                product_id = (
-                    self.product_template.product_variant_ids and
-                    self.product_template.product_variant_ids[0])
-                product_attributes = (
-                    product_id._get_product_attributes_values_dict())
-            else:
-                product_attributes = (
-                    self.product_template._get_product_attributes_inherit_dict(
-                        self.production_id.product_attributes))
-            self.name = product_id.name or self.product_template.name
-            self.product_uom = self.product_template.uom_id
-            self.product_id = product_id
-            self.product_attributes = product_attributes
-
-    @api.one
-    @api.onchange('product_attributes')
-    def onchange_product_attributes(self):
-        product_obj = self.env['product.product']
-        self.product_id = product_obj._product_find(self.product_template,
-                                                    self.product_attributes)
